requirements.txt
pyserial
pyqt6